Break apart schema into one class per step (#17)
* Change schema and convert biography pages
This is partially-working. Biography information is not actually saved. Committing and pushing for code review.

* Finish updating bands plugin for new schema
All pages tested as working. Refactor complete.

* Rename columns for less verbose variables
Now that we've split up the schema, we don't need things like 'band.bio.bio_pic_url'.

* Change wants_panel to ENUM from INT
The wants_panel Column used an odd workaround before which was unintuitive. A set of options is more idiomatic. Fixes #19.

* Use getattr to avoid CSV output failure
Fixes #21, albeit in a clunky way.

* Add Band completed() func for automated emails
Now that the "completed" property is underneath each step, and the step for a band doesn't exist until the band completes that step, we need a way for the automated email script to check if a band has completed a step without erroring. This adds that function.